+++ datastore-design/go-grpc-demo/storagebackend/timescaledb/timescaledb.go
@@ -3,7 +3,6 @@ package timescaledb
 import (
 	"database/sql"
 	"datastore/common"
-	"datastore/datastore"
 	"fmt"
 	"os/exec"
 
@@ -19,108 +18,6 @@ type TimescaleDB struct {
 // Description ... (see documentation in StorageBackend interface)
 func (sbe *TimescaleDB) Description() string {
 	return "TimescaleDB database"
-}
-
-// AddTimeSeries ... (see documentation in StorageBackend interface)
-func (sbe *TimescaleDB) AddTimeSeries(request *datastore.AddTSRequest) error {
-	var err error
-	var rows *sql.Rows
-
-    // check if time series ID already exists
-	rows, err = sbe.Db.Query("SELECT id FROM time_series WHERE id = $1", request.Id)
-	if err != nil {
-		return fmt.Errorf("sbe.Db.Query(1) failed: %v", err)
-	}
-	if rows.Next() {
-		return fmt.Errorf("time series ID %d already exists", request.Id)
-	}
-
-	// check if (station ID, param ID) combo already exists
-	rows, err = sbe.Db.Query(
-		"SELECT id FROM time_series WHERE station_id = $1 AND param_id = $2",
-		request.Metadata.StationId, request.Metadata.ParamId)
-	if err != nil {
-		return fmt.Errorf("sbe.Db.Query(2) failed: %v", err)
-	}
-	if rows.Next() {
-		var id int64
-		err = rows.Scan(&id)
-		if err != nil {
-			return fmt.Errorf("rows.Scan(&id) failed: %v", err)
-		}
-		return fmt.Errorf(
-			"(station_id, param_id) combo (%s, %s) already exists for time series ID %d",
-			request.Metadata.StationId, request.Metadata.ParamId, id)
-	}
-
-	// insert new time series
-    cmd := `
-	    INSERT INTO time_series (id, station_id, param_id, pos, other1, other2, other3)
-        VALUES ($1, $2, $3, ST_MakePoint($4, $5), $6, $7, $8)
-    `
-    _, err = sbe.Db.Exec(
-	    cmd, request.Id, request.Metadata.StationId, request.Metadata.ParamId, request.Metadata.Lon,
-		request.Metadata.Lat, request.Metadata.Other1, request.Metadata.Other2,
-		request.Metadata.Other3)
-	if err != nil {
-		return fmt.Errorf("sbe.Db.Exec() failed: %v", err)
-	}
-
-	return nil
-}
-
-// PutObservations ... (see documentation in StorageBackend interface)
-func (sbe *TimescaleDB) PutObservations(request *datastore.PutObsRequest) error {
-	// TODO
-	return fmt.Errorf("TimescaleDB/PutObservations() not implemented yet")
-}
-
-// GetObservations ... (see documentation in StorageBackend interface)
-func (sbe *TimescaleDB) GetObservations(request *datastore.GetObsRequest) (
-	*datastore.GetObsResponse, error) {
-
-	// TODO
-
-	// tsobs := []*datastore.TSObservations{}
-
-	// // --- BEGIN TODO: retrieve observations from database -----------
-	// // ... for now:
-	// var nSteps int64 = 3
-	// max := func(x, y int64) int64 {
-	// 	if x > y {
-	// 		return x
-	// 	}
-	// 	return y
-	// }
-	// timeStep := max(1, (request.Totime - request.Fromtime) / nSteps)
-	// for _, tsid := range request.Tsids {
-	// 	obs := []*datastore.Observation{}
-	// 	obsTime := request.Fromtime
-	// 	for i := 0; obsTime <= request.Totime; i++ {
-	// 		obsVal := 10 + float64(i)
-	// 		obs = append(obs, &datastore.Observation{
-	// 			Time: obsTime,
-	// 			Value: obsVal,
-	// 			Metadata: &datastore.ObsMetadata{
-	// 				Field1: fmt.Sprintf("value1 (%d)", i),
-	// 				Field2: fmt.Sprintf("value2 (%d)", i),
-	// 			},
-	// 		})
-	// 		obsTime += timeStep
-	// 	}
-	// 	tsobs = append(tsobs, &datastore.TSObservations{
-	// 		Tsid: tsid,
-	// 		Obs: obs,
-	// 	})
-	// }
-	// // --- END TODO: retrieve observations from storage backend -----------
-
-	// return &datastore.GetObsResponse{
-	// 	Status: -1, // for now
-	// 	Tsobs: tsobs,
-	// }, nil
-
-	return nil, fmt.Errorf("TimescaleDB/GetObservations() not implemented yet")
 }
 
 // openDB opens database identified by host/port/user/password/dbname.
